Added a whole bunch of is-dev checks to be able to push some new features to prod without pushing all of them, as well as adding more protections and QoL features to the checkpoints system

+++ src/main/java/com/denied403/hardcoursecheckpoints/Events/onQuit.java
@@ -20,7 +20,8 @@ public class onQuit implements Listener {
             sendMessage(e.getPlayer(), null, "leave", null, null);
         }
         Double highestLevel = database.getLevel(e.getPlayer().getUniqueId());
-        if(highestLevel <= 3){
+        int season = database.getSeason(e.getPlayer().getUniqueId());
+        if(highestLevel <= 3 && season == 1){
             database.deleteSpecific(e.getPlayer().getUniqueId());
         }
     }
